requirements.txt
# Core ML
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.45.0
diffusers>=0.31.0
accelerate>=1.0.0
safetensors>=0.4.0

# 3D Processing
trimesh>=4.4.0
pymeshlab>=2023.12
open3d>=0.18.0
xatlas>=0.0.9
pygltflib>=1.16.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
rembg>=2.0.0

# Evaluation Metrics
torchmetrics>=1.3.0
lpips>=0.1.4
pytorch-fid>=0.3.0

# Data Pipeline
huggingface-hub>=0.24.0
objaverse>=0.1.7
tqdm>=4.66.0
pandas>=2.2.0

# Training Utilities
tensorboard>=2.17.0
wandb>=0.17.0

# API Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Rendering
bpy>=4.0.0  # Blender Python (install separately)

# Utils
pyyaml>=6.0
requests>=2.32.0
numpy>=1.26.0
scipy>=1.14.0
